@@ -9,26 +9,24 @@
         "notification",
         "google"
     ],
-    "homepage": "http://packages.zendframework.com/",
+    "homepage": "https://github.com/zendframework/zendservice-google-gcm",
     "license": "BSD-3-Clause",
     "autoload": {
-        "psr-0": {
-            "ZendService\\Google\\Gcm\\": "library/",
-            "ZendService\\Google\\Exception\\": "library/"
+        "psr-4": {
+            "ZendService\\Google\\": "library/"
         }
     },
-    "repositories": [
-        {
-            "type": "composer",
-            "url": "http://packages.zendframework.com/"
+    "autoload-dev": {
+        "psr-4": {
+            "ZendServiceTest\\Google\\": "tests/"
         }
-    ],
+    },
     "require": {
-        "php": ">=5.3.3",
-        "zendframework/zend-http": ">=2.0.0",
-        "zendframework/zend-json": ">=2.0.0"
+        "php": "^5.5 || ^7.0",
+        "zendframework/zend-http": "^2.0",
+        "zendframework/zend-json": "^2.0"
     },
     "require-dev": {
-        "phpunit/PHPUnit": "3.7.*"
+        "phpunit/PHPUnit": "^4.8"
     }
 }
