@@ -1,15 +1,14 @@
 <?php
 /**
- * Zend Framework (http://framework.zend.com/)
+ * Zend Framework (http://framework.zend.com/).
  *
  * @link       http://github.com/zendframework/zf2 for the canonical source repository
+ *
  * @copyright  Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd New BSD License
+ *
  * @category   ZendService
- * @package    ZendService_Google
- * @subpackage Gcm
  */
-
 namespace ZendService\Google\Gcm;
 
 use ZendService\Google\Exception;
@@ -17,11 +16,9 @@ use ZendService\Google\Exception;
 /**
  * Google Cloud Messaging Response
  * This class parses out the response from
- * the Google Cloud Messaging API
+ * the Google Cloud Messaging API.
  *
  * @category   ZendService
- * @package    ZendService_Google
- * @subpackage Gcm
  */
 class Response
 {
@@ -94,12 +91,14 @@ class Response
     protected $response;
 
     /**
-     * Constructor
+     * Constructor.
      *
-     * @param string $response
+     * @param string  $response
      * @param Message $message
+     *
      * @return Response
-     * @throws Exception\ServerUnavailable
+     *
+     * @throws \ZendService\Google\Exception\InvalidArgumentException
      */
     public function __construct($response = null, Message $message = null)
     {
@@ -113,7 +112,7 @@ class Response
     }
 
     /**
-     * Get Message
+     * Get Message.
      *
      * @return Message
      */
@@ -123,19 +122,21 @@ class Response
     }
 
     /**
-     * Set Message
+     * Set Message.
      *
      * @param Message $message
+     *
      * @return Response
      */
     public function setMessage(Message $message)
     {
         $this->message = $message;
+
         return $this;
     }
 
     /**
-     * Get Response
+     * Get Response.
      *
      * @return array
      */
@@ -145,32 +146,32 @@ class Response
     }
 
     /**
-     * Set Response
+     * Set Response.
      *
      * @param array $response
+     *
      * @return Response
+     *
      * @throws Exception\InvalidArgumentException
      */
     public function setResponse(array $response)
     {
-        if (!isset($response['results']) ||
-            !isset($response['success']) ||
-            !isset($response['failure']) ||
-            !isset($response['canonical_ids']) ||
-            !isset($response['multicast_id'])) {
+        if (!isset($response['results'], $response['success'], $response['failure'], $response['canonical_ids'], $response['multicast_id'])) {
             throw new Exception\InvalidArgumentException('Response did not contain the proper fields');
         }
+
         $this->response = $response;
         $this->results = $response['results'];
         $this->cntSuccess = (int) $response['success'];
         $this->cntFailure = (int) $response['failure'];
         $this->cntCanonical = (int) $response['canonical_ids'];
         $this->id = (int) $response['multicast_id'];
+
         return $this;
     }
 
     /**
-     * Get Success Count
+     * Get Success Count.
      *
      * @return int
      */
@@ -180,7 +181,7 @@ class Response
     }
 
     /**
-     * Get Failure Count
+     * Get Failure Count.
      *
      * @return int
      */
@@ -190,7 +191,7 @@ class Response
     }
 
     /**
-     * Get Canonical Count
+     * Get Canonical Count.
      *
      * @return int
      */
@@ -200,15 +201,15 @@ class Response
     }
 
     /**
-     * Get Results
+     * Get Results.
      *
      * @return array multi dimensional array of:
-     *         NOTE: key is registration_id if the message is passed.
-     *         'registration_id' => array(
-     *             'message_id' => 'id',
-     *             'error' => 'error',
-     *             'registration_id' => 'id'
-     *          )
+     *               NOTE: key is registration_id if the message is passed.
+     *               'registration_id' => [
+     *               'message_id' => 'id',
+     *               'error' => 'error',
+     *               'registration_id' => 'id'
+     *               ]
      */
     public function getResults()
     {
@@ -216,25 +217,27 @@ class Response
     }
 
     /**
-     * Get Singular Result
+     * Get Singular Result.
      *
-     * @param  int   $flag one of the RESULT_* flags
+     * @param int $flag one of the RESULT_* flags
+     *
      * @return array singular array with keys being registration id
      *               value is the type of result
      */
     public function getResult($flag)
     {
-        $ret = array();
+        $ret = [];
         foreach ($this->correlate() as $k => $v) {
             if (isset($v[$flag])) {
                 $ret[$k] = $v[$flag];
             }
         }
+
         return $ret;
     }
 
     /**
-     * Correlate Message and Result
+     * Correlate Message and Result.
      *
      * @return array
      */
@@ -247,6 +250,7 @@ class Response
                 $results[$id] = array_shift($results);
             }
         }
+
         return $results;
     }
 }
